GdOpt is an abstraction for Gradient Descent optimization.
GD optimization is always performed following the same scheme: A step
direction and a step size is found. The resulting step is taken in the
corresponding iteration. This abstraction is provided by the GdOpt
module, which replaces the Opt module. In GdOpt, all optimizers (those
that existed before as well) share a common interface GdOpt.opt(...).
They define their own specific update rule based on an arbitrary state
that they define as well.

The random and rosenbrock evaluation files are removed because they are
deprecated since some time, anyway.

Signed-off-by: mirkobunse <mirko.bunse@cs.tu-dortmund.de>